+++ App/component/store/userStore.js
@@ -119,19 +119,86 @@ class UserStore {
     //onPress={() => UserStore.login("-s", "tkarnr78^@")}
     let tempdata = {};
     //bbs에서 데이터를 가져와서 firebase json 형식에서 flatlist하기 좋은 형식으로 키값을 JSON 안으로 넣는다.
-    await firebase
-      .database()
-      .ref("user/data/" + userid)
-      .once("value", (snap) => {
-        // json을 string으로 바꾸었다가 다시 json 형식으로 표준화
-        // 그냥하면 안됌
-        // alert(JSON.stringify(snap.val()));/
-        tempdata = JSON.parse(JSON.stringify(snap));
-        if (snap.val() != null && tempdata.g == userpassword)
-          this.user = tempdata;
-        else alert("없는 아이디이거나 비밀번호가 다릅니다.");
-      });
+    if (userid.length < 5 || userpassword.length < 5) {
+      alert("아이디와 비밀번호는 5자리 이상이어야합니다.");
+    } else {
+      await firebase
+        .database()
+        .ref("user/data/" + userid)
+        .once("value", (snap) => {
+          // json을 string으로 바꾸었다가 다시 json 형식으로 표준화
+          // 그냥하면 안됌
+          // alert(JSON.stringify(snap.val()));/
+          tempdata = JSON.parse(JSON.stringify(snap));
+          if (snap.val() != null && tempdata.g == userpassword) {
+            this.user = tempdata;
+            return true;
+          } else {
+            alert("없는 아이디이거나 비밀번호가 다릅니다.");
+            return false;
+          }
+        });
+    }
+    return false;
   }
+  /*
+      function loginFunc() {
+      //유저가 없는지 확인
+      if (name.length < 5 || password.length < 5)
+        alert("아이디와 비밀번호는 5자리 이상이어야합니다.");
+      else if (name == "" || password == "") {
+        alert("아이디 또는 비밀번호가 빈 칸입니다.");
+      } else {
+        //유저 정보가 없다면 회원가입
+        firebase
+          .database()
+          .ref("user/data/" + name)
+          .once("value", (snapshot) => {
+            if (snapshot.val() == null) {
+              setName(name);
+              setPassword(password);
+              navigation.navigate("홈", {
+                userkey: name,
+              });
+              const val = {
+                a: "지역",
+                b: "이메일",
+                d: 0,
+                e: new Date(),
+                f: name,
+                g: password,
+                h: name,
+                i: name,
+                j: "01000000000",
+                k: "image:url",
+                l: "소속",
+                n: 1,
+              };
+              firebase
+                .database()
+                .ref("user/data/" + name)
+                .set(val);
+              alert("회원 가입이 완료되었습니다.");
+            } else {
+              //유저 정보가 있다면 비밀번호 확인
+              if (password == snapshot.val().g) {
+                //패스워드랑 아이디가 맞다면 로그인
+                setName(name);
+                setPassword(password);
+                navigation.navigate("홈", {
+                  userkey: name,
+                });
+                alert("정상적으로 로그인되었습니다.");
+              } else {
+                alert("잘못된 비밀번호입니다.");
+              }
+            }
+          });
+      }
+
+      // const [clientPassword, onChangePassword] = React.useState(null);
+      // let newBbsKey = firebase.database().ref("user/data").push();
+}*/
 
   printUserStore() {
     alert(this.user);
